Remove setSearchPath, add setCS205searchpath
Due to how Postgres manages the search_path for a user, it was not
feasible to create a function that sets the search path for a user
dynamically. This is because any variable (to be used safely) must be
enclosed in either single or double quotes. This created an issue with
the SET search_path syntax, as it would set it to a schema literally
named "$user, public, _shelter, _pvfc", rather than to the four schemas
separately.

The setCS205SearchPath function was brought back and placed into a
separate file specifically for use with the cs205 course.