Replace NAME and TEXT with std compliant types
NAME -> VARCHAR(50)
TEXT -> VARCHAR(128)

NAME was used to refer to rolenames/identifiers, but using this type is
discouraged. Currently, the default for the maximum length of an identifier
is 63 characters. To have a "round" number that should be compatible with
most systems, it was replaced with VARCHAR(50)

TEXT was used wherever it was not necessary to place an upper limit on the
length of a string, such as with passwords. It is practically equivalent to
the standards compliant CLOB type. However, Postgres does not support or
provide a mapping to the CLOB type. In most cases, the TEXT type was used to hold a password, so 128 characters should be sufficient.