Add dropUser function
The dropUser function drops a user and removed their schema, regardless of
multiple memberships. It also checks to see if they are a Student and/or
Instructor, and removes their entries in the Student and/or Instructor tables
if necessary. A version of this function may eventually the other three drop
functions.