Create and implement a ClassDB role for management
Rather than set the owner of the functions in prepareClassDB.sql to
DBManager, they are now set to be owned by a newly created role: ClassDB
This increases the security and stability of the system.

Currently only the functions in prepareClassDB.sql are set to be owned
by ClassDB. Functions in other scripts should be updated as needed.

Resolves #27