Add foldPgID function to addHelperFunctions
foldPgID replicates Postgres's folding behavior, in that unquoted
idenfiers are made lowercase. Quoted identifiers have their outer quotes
removed.

The datatypes VARCHAR(65) and VARCHAR(63) help to show the intended
use and behavior of the function. Identifiers are limited to 63 chars,
but +2 is needed for a quoted identifier. Postgres does not actually
enforce these values.