Use foldPgID in prepareClassDB wherever needed
The new foldPgID function is used wherever a straight string match of a
role name would have failed. A couple examples would be with pg_has_role
and when checking to see if the role existed in pg_roles.

The foldPgID function currently has to be manually declared and the
classdb schema manually created since there is currently a circular
dependency between addHelperFunctions and prepareClassDB.