Implement RoleBase compatable GroupRole functions and views
This commit provides an initial (i.e. untested) implementation of the GroupRole-specific functions and views defined for M2.

This includes the Instructor, Student, and DBManager views, and the following functions defined for each group:

addClassDBRolesMgmt.sql:
- createXYZ
- revokeXYZ
- dropXYZ
- dropAllXYZ

addHelpers:
-isXYZ

This results in the replacement of all code that was in addClassDBRolesMgmt.sql (previously named addUserMgmt.sql)

The implementation for the three views is currently commented out as the User view they are based off of is not yet implemented