Alter isXYZ functions to also check if role is known
The isStudent, isInstructor, and isDBManager functions currently defined in addClassDBRolesMgmt.sql now also check that the role in question is known to ClassDB (i.e. an entry for that role exists in the RoleBase table)

Due to this change, the execution privileges for the three functions have been alters so that only Instructors and DBmanagers can execute them.