Semi-automate privilege tests
Running privilege tests no longer require the user to mannually reconnect to the database and type in a script name. This is now taken care of by a psql script. The readme has also been updated to reflect this.

However, the results of the tests still need to be manually checked.