Change test case to assign objects on drop
Change test case to assign objects ClassDB_Instructor, because objects can no longer be assigned to a ClassDB group role. That case now assigns object to to "current user".

This test case needs some extra work because of the restrictions on who has access to ClassDB schema and the dropRole function.