Add initial implementation of createTeam
Adds an untested implementation of ClassDB.createTeam(). This function cannot be used yet as the classdb_team group role has not been added at this point in branch history.

Additionally, issue #223 means that a temporary workaround is needed for the fullName of a team. Currently, a coalesce with teamName is used.