Add function and tests for dropping a team
Adds an untested function for dropping a team. The behavior is similar to other dropXYZ functions in this script. See Issue #220  for more information. Note that behavior in regards to pre-existing members is not yet investigated.

Tests knowingly fail, due to unmet dependencies in this branch (see earlier commits and related issues/PRs).