Guard use of pg_signal_backend
Use the pre-defined role pg_signal_backend only in server versions 9.6 or later. Fixes #225

Note: Depending on server version is technically not correct because the functionality in question is determined on the existence of a specific server role which can be tested in the pg_roles table. However, testing for server version is more practical and maintainable.